edits add front zero
This now adds zeros to represent the pedigrees in the ped file. for example paternal, maternal, sex, not used, and sample. 
Sample list must be a list of the samples in the CSV and the number of zeros in listofzeros must be the same as the number of samples in the csv